# springdoc-openapi-test-app2/api-definition.json
{"openapi":"3.0.1","info":{"title":"Petstore API","description":"This is a sample server Petstore server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters.","termsOfService":"http://swagger.io/terms/","license":{"name":"Apache 2.0","url":"http://springdoc.org"},"version":"1.1.44"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"device-api","description":"the device API"}],"paths":{"/device/{uuid}":{"get":{"tags":["device-api"],"summary":"Find device by uuid","description":"Returns a single device","operationId":"getDeviceById","parameters":[{"name":"uuid","in":"path","description":"uuid of pet to return","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"default response","content":{"application/xml":{"schema":{"$ref":"#/components/schemas/Device"}},"application/json":{"schema":{"$ref":"#/components/schemas/Device"}}}},"200":{"description":"successful operation","content":{"application/xml":{"schema":{"$ref":"#/components/schemas/Device"}},"application/json":{"schema":{"$ref":"#/components/schemas/Device"}}}},"404":{"description":"device not found","content":{"application/xml":{"schema":{"$ref":"#/components/schemas/Device"}},"application/json":{"schema":{"$ref":"#/components/schemas/Device"}}}}},"security":[{"XSPACE_TOKEN":[]}]}},"/device2/{uuid}":{"get":{"tags":["device-api"],"summary":"Find device by uuid","description":"Returns a single device","operationId":"getDeviceById2","parameters":[{"name":"uuid","in":"path","description":"uuid of pet to return","required":true,"schema":{"type":"string"}},{"name":"a_xyz","in":"query","required":true,"schema":{"type":"string"}},{"name":"c_ijk","in":"query","required":true,"schema":{"type":"string"}},{"name":"b_nmp","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"default response","content":{"application/xml":{"schema":{"$ref":"#/components/schemas/Device"}},"application/json":{"schema":{"$ref":"#/components/schemas/Device"}}}},"200":{"description":"successful operation","content":{"application/xml":{"schema":{"$ref":"#/components/schemas/Device"}},"application/json":{"schema":{"$ref":"#/components/schemas/Device"}}}},"404":{"description":"device not found","content":{"application/xml":{"schema":{"$ref":"#/components/schemas/Device"}},"application/json":{"schema":{"$ref":"#/components/schemas/Device"}}}}},"security":[{"api_key":[]}]}}},"components":{"schemas":{"Device":{"type":"object","properties":{"uuid":{"type":"string"},"name":{"type":"string"},"type":{"$ref":"#/components/schemas/DeviceType"}}},"DeviceType":{"type":"string","enum":["CONTAINER","POD","SERVICE"]}},"securitySchemes":{"XSPACE_TOKEN":{"type":"apiKey","name":"Auth_Token","in":"header"},"petstore_auth":{"type":"oauth2","flows":{"implicit":{"authorizationUrl":"http://petstore.swagger.io/oauth/dialog","scopes":{"write:pets":"modify pets in your account","read:pets":"read your pets"}}}}}}}